requirements.txt
aiohttp
botbuilder-core
botbuilder-dialogs
botbuilder-integration-aiohttp
python-dotenv
requests